+++ src/test/java/net/fabricmc/mappingio/read/DetectionTest.java
@@ -67,15 +67,21 @@ public class DetectionTest {
 	}
 
 	@Test
-	public void xrgFile() throws Exception {
+	public void xsrgFile() throws Exception {
 		MappingFormat format = MappingFormat.XSRG_FILE;
+		check(format);
+	}
+
+	@Test
+	public void jamFile() throws Exception {
+		MappingFormat format = MappingFormat.JAM_FILE;
 		check(format);
 	}
 
 	@Test
 	public void csrgFile() throws Exception {
 		MappingFormat format = MappingFormat.CSRG_FILE;
-		assertThrows(AssertionFailedError.class, () -> check(format));
+		check(format);
 	}
 
 	@Test
@@ -96,11 +102,24 @@ public class DetectionTest {
 		check(format);
 	}
 
+	@Test
+	public void recafSimpleFile() throws Exception {
+		MappingFormat format = MappingFormat.RECAF_SIMPLE_FILE;
+		assertThrows(AssertionFailedError.class, () -> check(format));
+	}
+
+	@Test
+	public void jobfFile() throws Exception {
+		MappingFormat format = MappingFormat.JOBF_FILE;
+		check(format);
+	}
+
 	private void check(MappingFormat format) throws Exception {
 		Path path = dir.resolve(TestHelper.getFileName(format));
 		assertEquals(format, MappingReader.detectFormat(path));
 
 		if (!format.hasSingleFile()) return;
+		if (format == MappingFormat.CSRG_FILE) return;
 
 		try (Reader reader = new InputStreamReader(Files.newInputStream(path), StandardCharsets.UTF_8)) {
 			assertEquals(format, MappingReader.detectFormat(reader));